+++ src/appleseed/renderer/meta/tests/test_bsdfmix.cpp
@@ -200,7 +200,7 @@ TEST_SUITE(Renderer_Modeling_BSDF_BSDFMix)
         ShadingPoint shading_point;
         ShadingPointBuilder builder(shading_point);
         builder.set_primitive_type(ShadingPoint::PrimitiveTriangle);
-        builder.set_uvs(Vector2d(0.0));
+        builder.set_uvs(Vector2f(0.0));
 
         BSDF& parent_bsdf = *assembly.bsdfs().get_by_name("parent_bsdf");
         parent_bsdf.evaluate_inputs(
